copy_on_clone - add machine applicability
